+++ qpsolvers/solve_ls.py
@@ -61,7 +61,7 @@ def solve_ls(
                 & lb \\leq x \\leq ub
         \\end{array}\\end{split}
 
-    using one of the available QP solvers.
+    using the QP solver selected by the ``solver`` keyword argument.
 
     Parameters
     ----------
@@ -87,7 +87,7 @@ def solve_ls(
         function. The standard L2 norm (W = Identity) is used by default.
     solver :
         Name of the QP solver, to choose in
-        :data:`qpsolvers.available_solvers`.
+        :data:`qpsolvers.available_solvers`. This argument is mandatory.
     initvals :
         Vector of initial `x` values used to warm-start the solver.
     sym_proj :